PYTHON:Refactors Brave Search connector structure
Reorganizes Brave Search connector by merging files into a single module for improved maintainability. Consolidates settings, response, and web page models. Updates imports and retains functionality.
