--- python/semantic_kernel/connectors/search/brave/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from semantic_kernel.connectors.search.brave.brave_search import BraveSearch
-from semantic_kernel.connectors.search.brave.brave_search_response import BraveSearchResponse
-from semantic_kernel.connectors.search.brave.brave_web_page import BraveWebPage
-
-__all__ = ["BraveSearch", "BraveSearchResponse", "BraveWebPage"]
